Seperate the charge version flagging from the overall service
